rename ssd1306 to monochromeoled
